pyproject.toml
[project]
name = "masim"
version = "0.1.0"
description = "Masim"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "fastapi>=0.120.0",
    "langchain[openai]>=1.0.0",
    "langgraph>=1.0.0",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langsmith>=0.4.37",
    "manim>=0.19.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "tzdata>=2025.2",
]

[dependency-groups]
dev = [
    "uvicorn[standard]>=0.38.0",
]
